Create the createGameBoard() function
Need to focus on making the tic-tac-toe
game work first on the console before DOM
manipulation, HTML and CSS.

The createGameBoard() is a factory function
and also an IIFE (module pattern). it defines
3 rows, 3 columns. It also defines board as an
empty array.

It then creates a 2D array with nested
for loops.
The outer for loop initializes at 0,
and for each time that i < rows (3), it creates
an array in board with an index of whatever
value that i is at that time. It then
initializes another for loop that that sets
j = 0. For each time that j < columns (3), it
pushes 0, into the array
that corresponds to the index of whatever
value that i is at that time (the array that
was just created inside board by the previous
loop). The nested loop increments and continues
this until j is no longer less than columns,
then it stops and goes back to the outer loop.
It continues that way until i is no longer less
than rows.

It pushes 0 for now but later on, it will push
a cell() function which will take care of the
cells in the board.

Also Create a function in createGameBoard()
then returns the board array.

In the HTML, add the script tag and a source
 attribute of the javascript file.