Create an addMarker() function
The function takes in 2 arguments
(row and column). The
function checks if the If the index of
the element of the row array
inside the board array (board[row][column])
is 0. For now, 0 represents empty cells.
If it is empty, it makes the element "x"
(for now) else it returns. This would
prevent players from adding markers to
spots that have already been taken.