Create createCell() and printBoard() method
First change the createGameBoard() from an IIFE
of module pattern to a normal factory function
for uniformity with the createCell() function.

With an IIFE, a const variable is not initialized
until its declaration at runtime so createCell()
will not be able be called before its declaration.

The createCell() defines a default cell value of 0.
This will make the default cell value anywhere
the function is called to be 0. It also
returns the value with the getCellValue() method.

In the createGameBoard(), the for loop pushes the
createCell() into the board array instead of 0.
This is because the createCell() will contain
useful methods for handling cell content.

Edit the addMarker() method such that the if
statement uses the getCellValue() method to check
if the value of the cell is 0 and if so, logs
something to the console (for now). Able to call
getCellValue() here because createCell() was
called in the for loop above and thus getCellValue()
has closure. Each createCell() retains access
to its properties and methods due to closure.

Define a printBoard() method that will go through
all the elements of board (rows), go through all
the elements of rows and use the getCellValue()
to get the value of all elements in each row.
This was done because getBoard() now console
log createCell() as it is createCell() that
is now being pushed into each cell.

Set the createGameBoard() to the gameBoard constant
and call the printBoard() method. Also call the
addMarker() method to verify if it logs the string
that was set.