Create the addPlayerMarker() method
The addPlayerMarker() method takes in
a player argument and sets the
value of the cell to value of the player
argument. This updates the internal state
of the cell to reflect that it now contains
a marker from the specified player.

Update the addMarker() method to use accept
one more argument - player. Now it updates
the cell (board[row][column]) with the
addPlayerMarker() method. Here the
addPlayerMarker() method is called with the
player argument that was passed in the addMarker()
method and updates the value of the cell to
the player's marker. A playRound() will be
created later to call the addMarker() method.