Create the controlGame() function.
The function would control the game.
It takes in two arguments with their
default values.

Assign the invocation of the
createGameBoard() function to the
gameBoard constant so its methods
can be called later.

Create a players array that two objects.
Each contains 2 properties - a name
and a marker. These properties will be
used to update the board later.