Get active player and print new round
Define a function getActivePlayer()
that returns the activePlayer.

Define another function printNewRound()
that calls the printBoard() method of the
createGameBoard() function. It then logs
a new round message that contains the name
property of the player who's turn it is.
It uses the dot notation to access the object
that is returned by getActivePlayer() -
getActivePlayer().name.
getActivePlayer() returns activePlayer and
activePlayer is an object with a name
property