Create playRound() method
This method will be used to play a round.
It accepts 2 arguments and first logs a
message that contains the row, column and
the name of the player.

It then calls the addMarker() method with
the it's own (playRound()'s) row and column
arguments. It also calls
the marker of the getActivePlayer() as the
3rd argument - addMarker(row, column, player)
This places the marker of the activePlayer
on the board.
It calls switchPlayerTurn() to switch players
and calls printNewRound() to print the updated
state of the board (with the new
marker) to the console.
Return playRound() and getActivePlayer().
Will use getActivePlayer() when building the
UI.

Assign the invocation of the controlGame()
function to the game constant and play 2
rounds.

Change the markers in the objects in the players
array to 1 and 2, to reflect the order of the
players. Will change to x and o later.