Print the board on game start
Call the printNewRound() method after
the playRound() so that when the game
starts, before any players have made
any moves, the board is printed on the
console and it is indicated that it is
player one's turn.

Edit the names of addMarker() and
addPlayerMarker() to placeMarker() and
placePlayerMarker() respectively.
Replace all instances of the old names
with the new ones.