Return the board
Create returnBoard() function to return
the board with values the same way
printBoard() did only that it does not log
it to the console.

Leave printBoard() because it logs
to the console to print state of board.
returnBoard() just returns board so that
functions can use it to determine a winner
and a draw situation.