Extract board values in threes
First return returnBoard() so that it
can be accessed by the getThreeValues()
function.

Create a getThreeValues function() that
extracts a sequence of three characters
from the returnBoard() array based on a
starting position and an offset. Later the
start and offset values that would target
all rows, columns and diagonals will be
plugged into the function. It will thus
extract all rows,columns and diagonals and
look for "xxx" or
"ooo" to determine if playerOne or playerTwo
won.

The initializes the result variable as an
empty string. The for loop runs 3 times and
concatenates the current cell value (based on
the specified index) to the
result string in each iteration of the loop.
It then returns result.