Check for winner
Change the player markers to "x" and "0"
because its better to work with strings
for now.

Create a checkForWinner() function that
checks for a winner.
The function initializes an empty array,
winningCombos. It then pushes all rows,
columns and diagonals with the getThree()
function into the empty array.

Check if any of the winning patterns, "xxx"
and "ooo" are found in the winningCombos
array and assign the results of those checks
to the playerOneWins and playerTwoWins
constants. Write an if else block that logs
the winner, depending on the winning pattern
found.

Call the checkForWinner() function in the
playRound() function.