Check for a tie
Initialize a playersMoves variable
in the controlGame() function scope
so that it can be accessed by both
the playRound() and the checkForTie()
A tic-tac-toe game results in a tie if
there are a total of 9 moves without a
winner.
Create the checkForTie() that logs
a message if playersMoves is 9.

In the playRound(), increment playersMoves
after each move (after each placeMarker())
so that it will result in a draw if it is
9 moves

Call checkForTie() in playRound().
Play 9 moves that results to a tie.