Edit checkForWinner()
Change the name of checkForWinner() to
checkWinnerOrTie() so that it can
handle checking for winnings and ties.
This is because the winningCombos array
needs to be accessed by a new tie
condition and want to leave the array
in the checkWinnerOrTie() scope.
Change the instance of checkForWinner() to
reflect the new checkWinnerOrTie().

Add the tie condition that is in the
checkForTie() function in another if
else block after the playerTwoWins
block.

Remove the checkForTie() function and
its instance.

Move the playersMoves instantiation to
be above the new checkWinnerOrTie()
instantiation.