Give the checkWinnerOrTie() a return value
Assign the tie condition to a constant
and replace the condition in
the tie's else if block with the constant.

Return the
playerOneWins or playerTwoWins or tie so
that the checkWinnerOrTie() returns true if
any of those values are truthy.

In the playRound(), Assign the invocation
of checkWinnerOrTie() to a constant so that
its return value can be used later.