Start setting up display
In the HTML body, create a parent
div .container and two children.
The .display-turn h1 would display
which player's turn it is to play
and the .board div would hold the
board

Return the getBoard method of
createGameBoard() in
controlGame() so that it can
be used to get the newest version
of the board after each player
move.

Create screenController(). Inside
it, assign the invocation of the
controlGame() to a constant.
Also set constants that target the
display-turn and board elements