Create an updateDisplay()
This function will update the display
with details of both player turn and the
board. It clears the board. It
then assigns the invocation of the
game.getBoard() (the returned getBoard
method of createGameBoard()) to a
constant. Wherever this is called,
it will get the newest
state of the board (each user move).

It also assigns the activePlayer to a
constant and sets the playerTurnDiv
to display the activePlayer's name.

In the CSS add some basic styles for
the board.