Render the board squares
With a nested forLoop, loop through
the board and for each element of the
board (row), get it index, loop through
its elements, get the index
of each of their indexed
a create a button for each
one of its elements.

Give it a class two data attributes -
one corresponds to its rowIndex and the
other, its colIndex. These attributes
will be used to determine the row and
column of the button that is clicked.

Set the textContent of each cell with the
getValue() (accessible due to closure -
createCell() is
passed in createGameBoard(),
createGameBoard() in
controlGame() and controlGame() in
screenController()
).

Append these buttons to the board div.

In the HTML, add the link tag and link
the CSS.