Create click handler function
Create the handleBoardClicks().
It will handle the clicks on the
buttons in the board.

The function first creates two constants
where the the data-row and data-column
(data attributes) values of whatever
button that was clicked on are stored.

The function is returned early if the
data-row and data-column of whatever
was clicked does not exist. This means
that another element or a gap in the
board was clicked.