Play move if cell is empty
Wrap the playMove() in a conditional
statement that checks if the text
content of the element that was clicked
on. The playMove is only called if the
textContent of the element is "0". This
means that a move has not been played
on it and prevents players from playing
on cells that have already been taken.