Reset board, activePlayer and playersMoves
Create a resetCell() in createCell() that
just resets the cellValue to 0. Return the
resetCell().
In createGameBoard(), create a resetBoard().
The resetBoard() loops through each element
of board (row), loops through each element
of each row (cell) and calls the resetCell()
on each of them. This resets their values.

In controlGame(), create a
resetActivePlayer() that resets the active
player to playerOne.

These resets will be later used to
to reset the board and active player after
wins and ties.

In controlGame() change the player markers
to 1 and 2 respectively. Change them in 136
and 137 so that the right winning patterns
can be searched for. The o's were a bit
confusing to read since the default cell
value is still 0.

In the CSS, give the .cell a big font size
to make them better readable.