Create a resetGame()
Move the declaration of the
playersMoves variable before the
creation of the resetGame(), because
it looks nicer there.

The resetGame(), calls the resetBoard(),
resetActivePlayer(), and changes the value
of playersMoves back to 0.

In playMove(), wrap the switching of
player turn and printing of new round
in an if else block such that if the
game is over, it resets the game, else
it switches player turn and prints new
round.

Later resetGame would be called by the
screenController() as now, resetting the
board when there is a win or draw does
not allow the users to see the move that
results in a move or draw before the board
is reset.