Add temporary reset button
In the HTML, add a simple button
that would be used to reset the game
and disable it in the javascript.js.

The plan is to enable the button when
the game is over. This is temporary
as later, a better UI will be
implemented.

The bigger picture is to make the
state of the game accessible by the
screenController(). Once that is
possible, the possibilities are endless.