Change the return value of checkWinOrTie()
Edit the function such that it returns a
string if there is a winner or a tie. If
there is neither winner nor tie, it
returns null.

Now playMove() assigns checkWinOrTie()
to the gameResult constant. It checks
for gameResult and returns gameResult.
After the IF ELSE block, it returns null
so that playMove() effectively returns
gameResult or null.

This way, playRound() will return the game
result (the return value of checkcWinOrTie()).
The screenController() can be able to
access the game result since playRound()
is called inside it.