Reset the game in screenController()
First expose resetGame in the returned
object for controlGame() So that
screenController can call it.

Initialize another gameResult constant
in the handleBoardClicks() scope so
and assign the return value of the
invocation of the playMove() to it.
The screen is then
updated. An IF statement now checks for
gameResult and changes the textContent
of the playerTurnDiv and enables the
reset button. Once the game ends, the
playerTurnDiv displays the gameResult
and the reset button becomes active.

Attach a click event listener to
resetButton so that onclick, it resets
the game, updates the display and
disables the resetButton.

The UI will hopefully be improved later.