Start handling three-wins-in-a-row.
In the HTML, add two divs with
respective classes that will display
player scores.

In the screenController(), create two
variables - playerOneStreak and
playerTwoStreak that will hold the player
scores. Assign the references of the two
divs created for player scores to
respective constants.

In the handleBoardClicks(), add IF ELSE
blocks such that if any player wins, his
score is updated and the other player's
score is reset to 0. If there is a tie,
both players' scores are reset to 0. It
will be easier to handle three wins in a
row from here.

Change the text content of the divs to
the respective player score.