Add reset round functionality.
In the HTML, add a new-round button.
The button will reset the game after
any player has had 3 wins-in-a-row.

In the JS, in screenController(), assign
the reference of the new-round button
to a resetRound constant and disable it.

In handleBoardClicks(), create an IF
block that checks if any of the players
has had 3-wins-in-a-row and if so,
disable the resetButton and enable this
new-round button.

Below the resetButton's event listener,
add a click event listener for this
resetRound button. On click the button
calls the resetGame() method, updates
the display, and resets the both players'
streak scores. It also updates the scores'
divs to the current reset scores - if it
it not updated here, it will hold the old
scores until there is a click on the board
(a new game starts). This is because
inside handleBoardClicks() is where it
gets its next update. It then disables
both the resetButton and the resetRound
buttons to allow for normal game-play.

Improve on UI later.