Make the checkWinnerOrTie() dynamic
The plan is to make the code dynamic in
case controlGame() is called with
parameters.

First make the checkWinnerOrTie() dynamic
by having it return an object containing
the result and the player name that wins
(if there is a win) and an object
containing just the tie result if its a
tie. This is so the playerTurnDiv can be
dynamically updated with the winner.

Expose the players array in the
controlGame() so it can be called by
screenController() to dynamically update
the player names in the player score divs.