Display winner dynamically
In handleBoardClicks() edit the IF ELSE
block so that it checks for the result
property of the gameResult object, and
if its value is "win", it updates the
playerTurnDiv with the value of the
winner property. Nest another IF ELSE
statement inside the winning condition
that checks if the value of the winner
property is playerOne name or
playerTwo's name and updates their
player streaks accordingly.

After the win condition, add a tie
condition.

Update the player streak divs with the
player names.