Fix typeError in handleBoardClicks()
gameResult returns null if there is no
winner or draw yet. The
gameResult.result === "win" check will
try to access the result property when
there is no winner yet and will return a
typeError since gameResult is null then.
It keeps returning the error until the
game ends.

Move the whole logic (for displaying wins,
ties and handling winner streak) inside
the first IF condition that checks for the
result (and disables the resetButton). This
will make the logic to only run if
gameResult is truthy.

Due to the way the logic was previously
structured, the player names did
not update in the player score divs until
the game ends but now it updates once there
is a click on the board (to start the game).