Better handle the display of scores
The scores were only being updated inside
handleBoardClicks() so the player names would
not show until there is a click on the board
(to start) a game. Make the updateDisplay()
to also display the scores. This makes it to
display the names and scores of the players
at the beginning every time it is called. It
is called in the screenController() at 306 so
it renders right from the start, before any
moves are made. Now the names are displayed
before any moves are made and the scores do
not just wait for the resetButton to be
clicked before they are displayed.

If the scored display is left for the
updateDisplay() alone, it will not update
after a game ends until the resetButton is
clicked. It will thus not display the
score 3 that makes a player to win a round
and this is bad for UX.

Also update the player names in the
resetRound() event handler so that when it
is clicked, the player names are updated.

Remove the default player name and score
from the HTML because as the JS handles
the name and scores better now, it has
become redundant.