Build the DOM in screenController()
The plan is to have a form with input
fields where users enter player names
and start the game. The users should
also be able to start a new game midway
an ongoing game. Will need to clear the
board and render the form. Thus the
player turn display, the board and the
scores display has to be built in
screenController();

In screenController(), create the
gameContainer, playerTurnDiv,
boardDiv, playerOneScore, playerTwoScore
and give them relevant classes.

Append the gameContainer to the body.
Append the playerTurnDiv,
boardDiv, playerOneScore, playerTwoScore
to the gameContainer.

Comment the container, turn display, board
and scores display out in the HTML, leaving
only the reset buttons. They will be
handled differently later.