Add parameters to screenController()
Tweak the screenController()
so that it accepts two parameters -
playerOneName and playerTwoName.
On the next line, use these 2
parameters to replace the previous
arguments on the controlGame(). This
is so that screenController() can be
called with the values of player one
and player two text inputs. Calling
the screenController with these
arguments will play the game with
the same arguments.