Create newGame()
The newGame() builds a form, receives
player names as input and plays the
game with the player names.

The newGame() first creates a formDiv and
a form and gives them classes. It also
creates a div, playerOneDiv for playerOne
and gives it a class. It creates a label,
playerOneLabel for playerOne, gives it
an attribute and a text content. It
creates an input element playerOneName
for playerOne and gives it two
relevant attributes. It carries out these
same tasks for playerTwo.

It then creates a startGame button with a
relevant attribute, class and text content.

It appends the formDiv to the body and the
form to the form div. It appends the
playerOne div to the form and playerOneLabel
and playerOneName to the playerOneDiv. It
appends everything about playerTwo after
playerOne. It then appends the startButton.

It attaches a click event listener to the
start button that removes the form-div class
from the formDiv and adds the form-div-hidden
class.
This new class has a style of display: none.
This hides the formDiv from the display and
removes it from the normal document flow (it
does not occupy any space). The even listener
then calls screenController(), with the values
of playerOneName and playerTwoName as its
arguments.

This hides the form div, renders the board and
plays the game with the player names.

In the CSS, add styles for the form-div and
form-div-hidden classes.