Start new game during gameplay
In screenController(), line 196 assign
the creation of a button element to a
newGameButton constant, give it a class
and a text content. In line 213, append
it to gameContainer.

In handleButtonClicks(), attach a click
event listener to newGameButton such that
when it is clicked, it changes the class
of gameContainer to a class that has a
display: none style in the CSS and calls
newGame().

Thus during gameplay, players can click on
the button and the form where player name
input is received will be rendered and they
can start a new game.

In the CSS, add the class selector and the
declarations that hides the gameContainer.