Remove gameContainer and formDiv onClick
Clicking the formDiv start game
button changes its class to a class
that hides it and renders the
gameContainer. If the newGame button
in the gameContainer is clicked, it
changes the gameContainer's class
to a class that hides it and renders
the formDiv.

The next time that the newGame button
is clicked, it creates another form.
(the previous one remains in the DOM
and is hidden by CSS)
The next time that the startButton is
clicked, it creates a new gameContainer
- the previous one is hidden by CSS
also. This creates multiple redundant
formDivs and gameContainers during
gameplay.

Since the formDiv is no longer needed
when startGame is clicked and it is
rebuilt by the DOM when it is needed,
it is best to remove it with
remove() instead of hiding it. Same
with gameContainer.

In the CSS, Remove the class selectors
and their declarations that were hiding
the displays of formDiv and gameContainer.