Show modal when there is gameResult
Add a dialog element in the HTML.
The dialog contains a div that will
display the winner and another that
holds the reset-game and new-round
buttons. This is so that whenever
there is a winner or a tire, the
dialog is shown, with the two buttons.

In the screenController, assign the
reference of this dialog element to
the infoDialog constant.
In handleBoardClick, show this modal
when there is a gameResult. In the click
event listeners for resetButton
and resetRound, close the dialog after
the various resets have happened.

This resets the game or the round and
closes the dialog (depending on the
button clicked).