Add new game button to the dialog
In the HTML add a new game button
in the dialog element with a
relevant class.

In the JS,
screenController() (line 199) update
the newGameButton class to correspond
with the class of the button added
in the dialog. This is so that these
classes can be used to target these
buttons.

Assign the nodelist of
all element with the new-game-button
class to a constant. Edit the
click event handler that was for the
newGameButton such that it now loops
through this nodelist and adds and
event handler to these buttons.

The click event handler still removes
the gameContainer and calls newGame().
It then closes infoDialog if the
infoDialog has an attribute of open.
This is to close the new game button
in the infoDialog after it is clicked
and a new game is started.