Fix incrementing score for playerTwo
As at the last commit, the
playerTwoStreak was not incrementing
if playerTwo wins. It was previously
incrementing but somehow the IF
block was altered in the cause of
development. The
gameResult.result === "win" closing
brace was not properly nesting the
incrementing logic. Also, the IF
block of the incrementing logic
for playerOne did not have open
and closing braces.

This made only playerOneStreak
to increment.

Nest the incrementing logic
properly and add the braces for the
IF block of the
incrementing logic for playerOne .