Style game title and input labels
In the JS, change the class name of
gameTitle to a better name and
style it accordingly in the CSS.
Give it a grid display and place
its content at the center.

Give the player divs a grid display
and justify its items center so that
the labels can be in the center.
Also give the player divs a gap and
increase their font size - this
increases the font size of the labels.