Add styles to the form-h1 and form-div
Increase the size of the form-h1 and
give it a margin bottom. The margin
bottom reduces the space between it
and the top of the form div.

Reduce the height of the form-div and
give it a border radius. The more the
height of the form div, the farther the
form-h1 is from it.

Give the text input and the form start
button a border-radius. Giving the text
input a border radius adds a default
black border to it. Restyle it later.