Wrap the form contents in a div
Previously, the gameTitle and formDiv
were appended to the body. When the
startGame button is clicked, it removes
only the formDiv. The gameTitle is
stuck to the body above the gameContainer
as screenController() is called.

In newGame() create a startPageContainer
div and give it a relevant class. From
line 410, append the startPageContainer to
the body and append the gameTitle and the
formDiv to the startPageContainer.

Remove the startPageContainer instead of
the formDiv when the startGame button is
clicked. Now, the gameTitle and the is
not appended to the top of the
gameContainer when screenController()
is called.