Wrap the board in a proper container
The plan is to have a header and a
game title and wrap them in divs so
that they can be styled better.

Create the boardHeader div, give it
a relevant class, create the gameTitle
hi, give it a relevant class and a
text-content. Create the boardContainer
div and give it a relevant class.

All the contents in this game point
where previously appended the the
gameContainer. From line 227,
append the boardHeader to the
gameContainer and append the gameTitle
to the boardHeader. Append the
boardContainer to the gameContainer
and append every other element to the
boardContainer. This creates a kind
of header and main sections and is
easier to style.

The click event listener on line 359
still removes the gameContainer so
all works as expected.