Style start page and game container
Now that the form sections and the
board sections are properly wrapped
in their respective containers,
these containers can be styled better.

Remove the grid, place-content and
height properties from the body and
use them to style the start page
container and the board container
so that the form and the board can
have a grid display and be placed
at the center.