Style the board header and button
Style the board-header div with a
display of flex and justify its
contents with a space-around
so that there can be equal space
around each element and at the
edges of the board-header container.

The board-h1 and the new-game-button
now have equal spaces between them
ane the edges of the board-header.

Style the new-game-button with an
align-self: center property and value
so that does not stretch to fill its
container vertically and is aligned
to the center instead. Also give it
a font-size and a padding.