Add styles to the board and cells
First move the class declarations
and styles of everything concerning
the board after class declarations
and styles concerning the form.
This is because the form comes before
the board in the flow.

Decrease the width and height of the
board. Also increase the gap so there
would be wider gutters between the
cells.

Add a border radius to the cells.