Fix pesky margins and paddings
There were pesky margins below the
board header div and the display turn
h2.

The margin and padding resets were in
the body selector. Move them to the
universal selector to properly reset
margins and paddings. Now margins and
paddings can be styled as needed.

Should have done this earlier.