Style the game-container
First remove the height property
and from the start-page-div and
board-container that were chained
and give the start-page-div a
height. This is done because the
common height of 90vh that is
applied to both the start-page-div
and board-container was making the
board-container to create too much
space it and the board-header.
The board-container was
trying to occupy 90% of the
viewport height.
Now the space between the board
header and the board-container can
be controlled using a number of
methods.

Give the game-container a grid
display, give it 3 row tracks (there
may be need for a footer later) and
give it a gap. The gap puts space
between the board-header and the
board container.

Give the board-header a padding to
add space between its contents and
its borders.