Style board container
Give the board container a display
of grid and a gap so that the space
between display turn element, the
board and the player scores elements
can be controlled.

Reduce the height and width of the
board to make some extra space for
the display turn element and the player
scores elements.