Style the winner info dialog
Due to the margin:0 reset that was
added to the universal selector, the
dialog when open rendered at the
extreme left.

Style the dialog (when its open) with
margin: auto so that
equal margins are created on all sides
of the dialog, effectively centering
it. Give it a min-width of 100vw so
that it spans 100% of its viewport
width including. (width: 100vw does not
span the dialog to touch the edges of
the document. It spans but does not touch
the edges). Also give the dialog a
text-align of center to align its
'contents' at the center for now. Will
center with a better method later.