Realign the contents of board header
The board-header-div had a display of
flex and its contents were aligned
horizontally. Styling the game container
with a grid display and dividing it
into row tracks made the board-header-div
wider. The contents tried to occupy
space as they should and this made
the board-h1 and the new-game-button
to be a bit dis-aligned as the board-h1
seemed closer to the top than the
new-game-button.

Styling the board header with
align-items: center, centers its
contents about the cross-axis,
realigning the board-h1 and the
new-game-button better horizontally.