Create a scores div and style
In the JS, in screenController(),
line 220, create a scoresDiv and
give it a relevant class name.
This div would contain the each
player's name, marker and score
in sets of containers.

Append the scoresDiv to
gameContainer and append
playerOneScore and playerTwoScore
to the scoresDiv.

In the CSS, style the scores-div
with a flex display and justify its
content with space-between so there
can be space between the players
score divs.