Setup playerOne's score div
In the JS, in screenController(),
from line 223, create playerOneScoreDiv,
playerOneMarker, playerOneMoniker divs
and give them relevant class names.

Give playerOneMarker a text-content of x
and playerOneMoniker a text-content of
playerone's name.

Append playerOneScoreDiv to the scoresDiv
and append playerOneScoreDiv,
playerOneMarker and playerOneMoniker to
playerOneScoreDiv.

Now the playerOneScoreDiv contains
playerOne's marker, name and score.

Remove playerOne's name from
updateDisplay() on 298, handleBoardClicks()
0n 347 and in the resetRound button click
event listener on 365. This is so that all the
playerOneScore is rendered without the
playerone's name because it is already
rendered in another div.