Edit controlGame() arguments
In screenController() on line 191,
edit the arguments of controlGame()
so that they are strings instead of
parameters of screenController.

This is so that the game can be played
temporarily with hardcoded strings
as the players' name. This is to
enable styling the board without
always typing player names in
start-page form. Will return to
previous state later.

Also edit the class names of
playerOneScore and playerTwoScore
as a . (dot) was mistakenly put
before the class names.