Style players' containers
Style the player-one-container
and player-two-container with a
display of grid, a gap and
justify its items center. Give
them left and right padding so
there can be space on the left
and right of these containers

Style the markers with a font-size
Style the names with and the scores
with a font-size and font-weight
such that the names and the markers
look bigger than the marker.