Setup score board display on dialog
Set up the dialog to display the
player names and their respective
scores whenever it is shown.

In the HTML, create a
dialog-display-scores div element.
This element contains divs to
display player one and player two
names with spans inside these
divs that display name to display the
scores. In between these divs that
display names is a div that displays
"vs". This structure is setup for a
score board display like
"Player One 0 vs 1 Player Two".

Later JS will be used to update these
elements dynamically.