Update the dialog score dynamically
Assign the reference of the elements
created for name display and score
display for player one and player two
to respective constants.

Create text nodes with the player names
This because the spans are
inside the divs and using textContent
to update the div will overwrite the
spans and their contents.

For playerOne, the player one name in
the text node has a space after it
so that there will be a space between
it and the score. Use insertBefore()
to append this text node before
the span.

For playerTwo, use appendChild()
to append the text node because
appendChild appends after the
last child (the span) and that is
the behavior that is needed, given
the expected structure of the
scoreboard.

Update the scores of each player
in handleBoardClicks(), from
line 385. The score of player two
has a space after it, so that there
will be a space between the score
the player name given
the expected structure of the
scoreboard.