Update displayInfo when playerStreak is 3
Update handleBoardClicks() such that at
the point when a winner has been found and
the streak of the winner is being
incremented, if the streak is 3,
displayInfo is updated with a message.

A streak of 3 means that the player has won
the game and previously, even when the streak
is 3, displayInfo would display the default
message of a win.

Now the players can be see a different message
on the infoDialog when the game is over.