Update infoDialog buttons
In the HTML, update the Reset Game button
text to Next Round and change the class
name appropriately. In the JS, on line 292,
update the name of its reference and the
class selector appropriately. Rename all
instances of the resetGame(previous name)
to the new name.This is because after a
win or a draw, the players proceed to the
'Next Round', not a 'Reset Game'

In the HTML, update the new-round button
class name appropriately. In the JS, on
line 295,update the name of its reference
and the class selector appropriately.
Rename all its instances of the reference
to its current name.

In the HTML, add an id attribute to the
new-game-button so that it can be styled
differently from the new game button in
the boardHeader. Cannot change the class
because a nodelist is using the class
selector (a click event listener was added
to the two buttons with the .new-game-button
class name). id has a higher specificity
than class so it will be styled with
the id.