Style dialog-container and its contents
Style the dialog-container a with a grid
display and a gap so that the space
between its contents can be controlled.
Also give it a top and bottom padding.

Increase the font-sizes of display-info
and the dialog-display-scores. Give the
dialog-display-scores a bold font-weight
so it stands out.

Style the modal-buttons with a flex
display and justify its content center.
Give it a gap to control the space
between the buttons.