Update cell style by player marker
In handleBoardClicks() from line 349
assign the nodelist of all cells
(.cell) to a constant. With this
constant, loop through the cells
and update the class of each cell
by its marker. Later these cells
will be styled in the CSS, by their
markers (using the classes assigned
here)

In the players array in
controlGame(), update the player
markers accordingly.

In checkWinnerOrTie() update the
playerOneWins and playerTwoWins
condition so that they can
correspond with the players'
markers that were updated in
controlGame().