Add more fonts
Remove the raleway font from the
directory and the CSS. Add Fredoka,
Nunito and M PLUS Rounded 1c fonts.

Import these fonts to the CSS. Style
the body with M PLUS Rounded 1c.
From line 148, Style the form-h1
and board-h1 with Nunito. On 183,
replace raleway with Fredoka so
that form labels, the cells and
the players' markers
(in the scores-div) are styled with
Fredoka. Fredoka is more rounded
and gives the x and o better rounded
corners.