Adjust size of board
From line 172, use clamp() to
adjust width and height of the
.board to make page a bit responsive.
Clamp() ensures that the size of
.board remains responsive and adapts
to the size of its container
(.board-container) while
respecting the defined minimum and
maximum constraints.

The width and height of .board will
shrink as its container shrinks but
will never go beyond 100px by 100px.
The width and height of .board will
also expand with its container but
will never go above 400px by 400px.

Reduce the gap so that the gutters
between cells can be reduced and the
cells have have a bigger size.

Style all buttons with
M PLUS Rounded 1c font family and
fall back fonts.