Adjust size of form-div
On line 37, give input a width of
100% else the form-div will not shrink
and the clamp function used on its
with property will be for nothing.
The input will not shrink because
form-div is trying to keep the form
input to respect the font-size to
the font-size set for it (line 70).
Setting the width of the input to
100% ensures that they take up the
full width of their container,
allowing the container to resize
responsively.

Give the start-page-container a
padding to create space between
its contents. This because the edges
of the form-container were touching
the edges of the document when the
document is shrunk to a smaller
size.